pyproject.toml
[project]
name = "con2phys-to-nwb"
version = "0.1.0"
description = "Convert con2phys electrophysiology data to NWB format"
license = "BSD-3-Clause"
requires-python = ">=3.12"
dependencies = [
    "boto3",
    "pynwb>=2.8",
    "h5py",
    "numpy",
    "pandas",
    "pyyaml",
    "tqdm",
    "nwbinspector",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/con2phys_to_nwb"]
